document_service/pyproject.toml
# pyproject.toml

[project]
name = "document-service"
version = "0.1.0"
description = "A microservice to process and query documents."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [ 
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "sqlmodel>=0.0.18",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "celery>=5.4.0",
    "redis>=5.0.4",
    "pypdf>=4.2.0",
    "python-docx>=1.1.2",
    "sentence-transformers>=2.7.0",
    "chromadb>=0.4.24",
] 
[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "ruff>=0.4.4",
]
[tool.setuptools]
packages = ["app"]